+++ src/Module/Api/Method/Api4/Flag4Method.php
@@ -64,7 +64,7 @@ final class Flag4Method
             return false;
         }
         ob_end_clean();
-        $type      = ObjectTypeToClassNameMapper::map((string)$input['type']);
+        $type      = (string) $input['type'];;
         $object_id = $input['id'];
         $flag      = (bool)($input['flag'] ?? false);
         $user      = User::get_from_username(Session::username($input['auth']));
@@ -82,7 +82,8 @@ final class Flag4Method
         if (!InterfaceImplementationChecker::is_library_item($type) || !$object_id) {
             Api4::message('error', T_('Wrong library item type'), '401', $input['api_format']);
         } else {
-            $item = new $type($object_id);
+            $className = ObjectTypeToClassNameMapper::map($type);
+            $item      = new $className($object_id);
             if (!$item->id) {
                 Api4::message('error', T_('Library item not found'), '404', $input['api_format']);